requirements.txt
fastapi
uvicorn
python-multipart
streamlit
requests
pillow
matplotlib
opencv-python
numpy
scikit-learn
torch
torchvision
pydantic
pyyaml
pandas
seaborn
tqdm
rich